+++ lib/facebooker2.rb
@@ -48,3 +48,4 @@ require "facebooker2/rails/helpers/javascript"
 require "facebooker2/rails/helpers/request_forms"
 require "facebooker2/rails/helpers/user"
 require "facebooker2/rails/helpers"
+require "facebooker2/rack/post_canvas"
