--- lib/facebooker2/rack/PostCanvas.rb
@@ -1,19 +1,0 @@
-module Rack
-  class PostCanvas
-    
-    def initialize(app)
-      @app = app
-    end
-
-    def call(env)
-      request = Request.new(env)
-      
-      if request.POST['signed_request']
-        env["REQUEST_METHOD"] = 'GET'
-      end
-      
-      return @app.call(env)
-    end
-  
-  end
-end
